.Se corrijen las clases Manager, Builder ,Transaction y Block para corregir errores de conversion debido a utilizar Size_t 	* Se aplica refactor en clase Block para que nonce y bits sean 	del tipo pedido por enunciado: Unsigned Int y para corregir 	problemas de conversion 	* Se modifica el metodo de concatenar provisto por la funcion 	GetConcatenatedTransactions para evitar errores de conversion 	* Se corrije BlockChainBuilder::dificultad para evitar que 	cuente extra 	* Se crea metodo para convertir flotantes en la clase 	  Transaction

+++ Src/BlockChainManager.cpp
@@ -95,7 +95,7 @@ void BlockChainManager::proccesStatus(status_t status){
 
 
 #define DIFFICULTY_DEFAULT_VALUE 3
-size_t BlockChainManager::userDefinedDifficulty = DIFFICULTY_DEFAULT_VALUE;
+unsigned int BlockChainManager::userDefinedDifficulty = DIFFICULTY_DEFAULT_VALUE;
 
 void BlockChainManager::setUserDefinedDifficulty(int d){
 	if( d < 0 ){
@@ -103,7 +103,7 @@ void BlockChainManager::setUserDefinedDifficulty(int d){
 		std::cerr << "Error de Formato: Dificultad debe ser mayor a cero" << std::endl;
 		std::abort();
 	}
-	userDefinedDifficulty = (size_t) d;
+	userDefinedDifficulty = (unsigned int) d;
 }
